Convert Integer to binary new way

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -6,7 +6,7 @@
  */
 void print_binary(unsigned long int n)
 {
-	unsigned long int max = 1, b = 1, num = 0;
+	unsigned long long int max = 1, b = 1, num = 0;
 
 	while (n > max)
 	{
